app/src/main/java/com/jesusbadenas/kotlin_clean_architecture_project/viewmodel/UserDetailsViewModel.kt
package com.jesusbadenas.kotlin_clean_architecture_project.viewmodel

import androidx.lifecycle.LiveData
import androidx.lifecycle.Transformations
import com.jesusbadenas.kotlin_clean_architecture_project.common.BaseViewModel
import com.jesusbadenas.kotlin_clean_architecture_project.domain.repositories.UserRepository
import com.jesusbadenas.kotlin_clean_architecture_project.entities.User
import com.jesusbadenas.kotlin_clean_architecture_project.entities.mappers.UserDataMapper

class UserDetailsViewModel(
    userId: Int,
    userRepository: UserRepository,
    private val userDataMapper: UserDataMapper
) : BaseViewModel() {

    val user: LiveData<User> = Transformations.map(userRepository.user(userId)) { userData ->
        userDataMapper.mapFrom(userData)
    }
}
